requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-crispy-forms==2.0
django-import-export==3.2.0
et-xmlfile==1.1.0
fontawesomefree==5.13.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
Pygments==2.15.1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
